Refactor every package from "cuff" to "bjbites", module from "cuff-springboot" to "bjbites-springboot", and the main backend class to BJBitesSpringbootApplication.java
