+++ backend/src/main/java/bjbites/bjbites_springboot/BJBitesSpringbootApplication.java
@@ -1,0 +1,13 @@
+package bjbites.bjbites_springboot;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BJBitesSpringbootApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BJBitesSpringbootApplication.class, args);
+	}
+
+}
