+++ backend/src/main/java/bjbites/bjbites_springboot/controller/BuildingController.java
@@ -1,7 +1,7 @@
-package cuff.cuff_springboot.controller;
+package bjbites.bjbites_springboot.controller;
 
-import cuff.cuff_springboot.entity.Building;
-import cuff.cuff_springboot.repository.BuildingRepository;
+import bjbites.bjbites_springboot.entity.Building;
+import bjbites.bjbites_springboot.repository.BuildingRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
