+++ backend/src/main/java/bjbites/bjbites_springboot/controller/FoodTypeController.java
@@ -1,7 +1,7 @@
-package cuff.cuff_springboot.controller;
+package bjbites.bjbites_springboot.controller;
 
-import cuff.cuff_springboot.entity.FoodType;
-import cuff.cuff_springboot.repository.FoodTypeRepository;
+import bjbites.bjbites_springboot.entity.FoodType;
+import bjbites.bjbites_springboot.repository.FoodTypeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
