+++ backend/src/main/java/bjbites/bjbites_springboot/controller/UserController.java
@@ -1,7 +1,7 @@
-package cuff.cuff_springboot.controller;
+package bjbites.bjbites_springboot.controller;
 
-import cuff.cuff_springboot.entity.User;
-import cuff.cuff_springboot.repository.UserRepository;
+import bjbites.bjbites_springboot.entity.User;
+import bjbites.bjbites_springboot.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
