+++ backend/src/main/java/bjbites/bjbites_springboot/entity/User.java
@@ -1,4 +1,4 @@
-package cuff.cuff_springboot.entity;
+package bjbites.bjbites_springboot.entity;
 
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
