+++ backend/src/main/java/bjbites/bjbites_springboot/repository/BuildingRepository.java
@@ -1,6 +1,6 @@
-package cuff.cuff_springboot.repository;
+package bjbites.bjbites_springboot.repository;
 
-import cuff.cuff_springboot.entity.Building;
+import bjbites.bjbites_springboot.entity.Building;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BuildingRepository extends JpaRepository<Building, Integer> {
