+++ backend/src/main/java/bjbites/bjbites_springboot/repository/DietaryOptionRepository.java
@@ -1,6 +1,6 @@
-package cuff.cuff_springboot.repository;
+package bjbites.bjbites_springboot.repository;
 
-import cuff.cuff_springboot.entity.DietaryOption;
+import bjbites.bjbites_springboot.entity.DietaryOption;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DietaryOptionRepository extends JpaRepository<DietaryOption, Integer> {
