+++ backend/src/main/java/bjbites/bjbites_springboot/repository/FoodTypeRepository.java
@@ -1,6 +1,6 @@
-package cuff.cuff_springboot.repository;
+package bjbites.bjbites_springboot.repository;
 
-import cuff.cuff_springboot.entity.FoodType;
+import bjbites.bjbites_springboot.entity.FoodType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FoodTypeRepository extends JpaRepository<FoodType, Integer> {
