+++ backend/src/main/java/bjbites/bjbites_springboot/repository/PostRepository.java
@@ -1,10 +1,9 @@
-package cuff.cuff_springboot.repository;
+package bjbites.bjbites_springboot.repository;
 
-import cuff.cuff_springboot.entity.Post;
+import bjbites.bjbites_springboot.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface PostRepository extends JpaRepository<Post, Integer> {
