+++ backend/src/main/java/bjbites/bjbites_springboot/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package cuff.cuff_springboot.repository;
+package bjbites.bjbites_springboot.repository;
 
-import cuff.cuff_springboot.entity.User;
+import bjbites.bjbites_springboot.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
