--- backend/src/main/java/cuff/cuff_springboot/CuffSpringbootApplication.java
@@ -1,13 +1,0 @@
-package cuff.cuff_springboot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CuffSpringbootApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CuffSpringbootApplication.class, args);
-	}
-
-}
