+++ backend/src/test/java/bjbites/bjbites_springboot/CuffSpringbootApplicationTests.java
@@ -1,0 +1,13 @@
+package bjbites.bjbites_springboot;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CuffSpringbootApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
