--- backend/src/test/java/cuff/cuff_springboot/CuffSpringbootApplicationTests.java
@@ -1,13 +1,0 @@
-package cuff.cuff_springboot;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CuffSpringbootApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
